Remove soname patch as the upstream already sets it
The soname.patch was added directly in preparation of the Debian
packaging on top of the first imported upstream version.
Upstream changed and modified the variable 'version' from '0.1' to
'0.0.0' since the import of 0~git20190625+cb21bb3 (the first imported
version) but the patch wasn't turned back.

So currently the Debian packaging is adding a version 0.1.0 to the
created shared library files while upstream is still using 0.0.0. I
think it's not correct to still use the Debian added version and the
patch should be dropped.
OTOH the library isn't real versioned nor is using a stable ABI/API.